Conflicts:
	include/maidsafe/data_types/immutable_data.h
	include/maidsafe/detail/data_type_values.h
	src/maidsafe/data_store/tests/data_store_test.cc
	src/maidsafe/data_types/immutable_data.cc
	src/maidsafe/data_types/mutable_data.cc
	src/maidsafe/data_types/tests/data_types_test.cc

+++ src/maidsafe/data_store/data_buffer.cc
@@ -412,7 +412,7 @@ void DataBuffer::SetMaxDiskUsage(DiskUsage max_disk_usage) {
 
 fs::path DataBuffer::GetFilename(const KeyType& key) {
   return kDiskBuffer_ / (EncodeToBase32(boost::apply_visitor(get_identity_, key))
-            + boost::lexical_cast<std::string>(boost::apply_visitor(get_tag_, key)));
+        + boost::lexical_cast<std::string>(static_cast<int>(boost::apply_visitor(get_tag_, key))));
 }
 
 template<typename T>
